# src/docs/swagger.yaml
swagger: "2.0"
info:
  description: "Insert description here"
  version: "0.0.0.1ht"
  title: "PIGGRAM"
  termsOfService: "Insert link to term of services"
  contact:
    email: "Insert email here"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:3000"
...
schemes:
  - "https"
  - "http"
paths:
  /user/:
    post:
      tags:
        - "user"
      summary: "Creates a new account and logs in"
      description: ""
      operationId: "signup"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "User information"
          required: true
          schema:
            $ref: "#/definitions/UserSignup"
      responses:
        "200":
          description: "User created"
        "400":
          description: "Invalid input"
      security:
        - piggram_auth:
            - "write:users"
            - "read:users"
...
  /user:
    get:
      tags:
        - "user"
      summary: "Get the session user"
      description: ""
      operationId: "getUser"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        200:
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/definitions/User"
        400:
          description: "Invalid input"
        401:
          description: "unauthenticated"
        403:
          description: "Forbidden"
      security:
        - piggram_auth:
            - "write:users"
            - "read:users"
...
securityDefinitions:
  piggram_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:users: "modify user information"
      read:users: "read user information"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  User:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      username:
        type: "string"
      name:
        type: "string"
      description:
        type: "string"
      image:
        type: "string"
...
  UserSignup:
    type: "object"
    properties:
      username:
        type: "string"
      name:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      image:
        type: "string"
...